+++ R/intensityplot.R
@@ -32,6 +32,14 @@
 #' on the y-axis? Default is `TRUE`. Can also be a logical vector if some but
 #' not all plot components should have minimum and maximum values on the y-axis.
 #' Ignored for TextGrid component.
+#' @param highlight Named list giving parameters for differential
+#' highlighting of the intensity contour based on the time domain. This list
+#' should contain information about which parts of the plot to highlight, either
+#' done with the `start` and `end` arguments which must be numbers or numeric
+#' vectors, or using the `tier` and `label` arguments to highlight based on
+#' information in a plotted TextGrid. Further contains the optional arguments
+#' `color` (string or vector of strings, see `color`) and
+#' `drawSize` (integer).
 #' @param axisLabel String giving the name of the label to print along
 #' the y-axis when plotting intensity. Default is `Intensity (dB)`.
 #' @param drawSize Number indicating the line width of the intensity contour.
@@ -49,6 +57,7 @@
 intensityplot <- function(it, start, end, tfrom0=TRUE, tgbool=FALSE, lines=NULL,
                           focusTierColor='black', focusTierLineType='dotted',
                           range=NULL, color='black', ind=NULL, min_max_only=TRUE,
+                          highlight=NULL,
                           axisLabel='Intensity (dB)', drawSize=1) {
 
   if (!min_max_only[ind]) {
@@ -69,6 +78,18 @@ intensityplot <- function(it, start, end, tfrom0=TRUE, tgbool=FALSE, lines=NULL,
     start <- 0
   }
 
+  if (!is.null(highlight)) {
+    highlight_t <- c()
+    highlight_i <- c()
+    for (int in 1:length(highlight$start)) {
+      times <- which(it$t > highlight$start[int] & it$t < highlight$end[int])
+      highlight_t <- c(highlight_t, it$t[times], max(it$t[times] + 0.0001))
+      highlight_i <- c(highlight_i, it$i[times], NA)
+    }
+    if (!'color' %in% names(highlight)) highlight$color <- color
+    if (!'drawSize' %in% names(highlight)) highlight$drawSize <- drawSize
+  }
+
   plot(it$t, it$i, xlim=c(start, end+start), xaxt='n',
        ylim=range, yaxt=yax, type='l', col=color, lwd=drawSize)
   if (!min_max_only[ind] & ind != 1) graphics::axis(2, at=ytix)
@@ -80,6 +101,11 @@ intensityplot <- function(it, start, end, tfrom0=TRUE, tgbool=FALSE, lines=NULL,
                        lty=focusTierLineType[i])
     }
   }
+
+  if (!is.null(highlight)) graphics::lines(highlight_t, highlight_i,
+                                           col=highlight$color,
+                                           lwd=highlight$drawSize)
+
   graphics::mtext(axisLabel, side=2, line=3.5, cex=0.8)
 
 }